Add skill learning for adventurer, skill learning event handling and IA basic fight

+++ src/ChickenAPI.Data/Skills/SkillDto.cs
@@ -14,7 +14,7 @@
 
         public byte Class { get; set; }
 
-        public short Cooldown { get; set; }
+        public int Cooldown { get; set; }
 
         public byte CpCost { get; set; }
